backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
cryptography>=42.0.0
httpx>=0.26.0
python-multipart>=0.0.6
pycryptodome>=3.19.0
python-dotenv>=1.0.0
